Added loc and git hash to particle track save
I can't figure out how to save a string to a netCDF4 variable, so I saved the empty variables 'loc' and 'git_hash' which in their 'long_name' metadata have their descriptions and values. loc has the location of the model output used to move drifters, and git_hash has the unique identifier for the commit version of the code used to do the tracking.