Added interpolate function to tools.py
Have moved interpolation functions into a separate function, interpolate, in tools.py which contains options for moving between grid space, projected xy space, and lon/lat. It also has functions for moving from grid space to xy or lon/lat (but not the other way since it depends on the uniformity of the grid).