Added conditional statement to real z calculation
Added if statement to the end of the loop in run.py to control whether or not to bother calculating the real z coordinate for the drifters. In the case of surface-only drifters, I suspect that no one cares what the real z depth is as the drifters bob up and down on the surface.