Miscellaneous mostly minor changes
Mostly did some changes to being able to take in the aggregated netCDF model output appropriately in inout.py functions. Also added in a line to several functions in there as a workaround to an issue with netCDF thredds files input/output. Updated the manual notebook and pdf and eliminated some extraneous lines.