Changed tabs to spaces, updated manual
All of the tabs in the python scripts were converted to spaces. A note was added to the README with a helpful tip about compiling the code. tracpy is no longer imported in the manual notebook since the individual files can be imported directly since they are sitting in the same directory.