Added usebasemap parameter, rearranged: Manual works with TracPy class
Added `usebasemap` parameter so that pyproj is normally used in the grid projections since it is much faster than the basemap version, but the parameter allows for the option of using basemap instead if you know you want to plot with the TracPy object. I changed the order of things a little and some notational type stuff, and now the manual ipython notebook works. So, I think I will merge with the master branch now or soon!